[Dataset] Update utils and remove some unused code (#149)
* Update data & datasets

* Fix data __init__

* Fix gcnmix test
