+++ cogdl/data/__init__.py
@@ -1,20 +1,11 @@
 from .data import Data
 from .batch import Batch
 from .dataset import Dataset
-from .dataloader import DataLoader, DataListLoader, DenseDataLoader
-from .download import download_url
-from .extract import extract_tar, extract_zip, extract_bz2, extract_gz
+from .dataloader import DataLoader
 
 __all__ = [
     "Data",
     "Batch",
     "Dataset",
     "DataLoader",
-    "DataListLoader",
-    "DenseDataLoader",
-    "download_url",
-    "extract_tar",
-    "extract_zip",
-    "extract_bz2",
-    "extract_gz",
 ]
