--- cogdl/data/makedirs.py
@@ -1,11 +1,0 @@
-import os
-import os.path as osp
-import errno
-
-
-def makedirs(path):
-    try:
-        os.makedirs(osp.expanduser(osp.normpath(path)))
-    except OSError as e:
-        if e.errno != errno.EEXIST and osp.isdir(path):
-            raise e
