+++ cogdl/datasets/matlab_matrix.py
@@ -1,13 +1,11 @@
-import json
-import os
 import os.path as osp
 from itertools import product
 
-import numpy as np
 import scipy.io
 import torch
 
-from cogdl.data import Data, Dataset, download_url
+from cogdl.data import Data, Dataset
+from cogdl.utils import download_url
 
 from . import register_dataset
