+++ cogdl/models/emb/knowledge_base.py
@@ -4,8 +4,6 @@ from __future__ import print_function
 
 import logging
 
-import numpy as np
-
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
